eval/safety-eval-fork/requirements.txt
alpaca-eval>=0.6.2
beaker-gantry>=0.22.2
beaker-py>=1.26.5
cloudpickle>=3.0.0
datasets>=2.19.0
dill>=0.3.8
fastchat>=0.1.0
fire>=0.6.0
huggingface-hub>=0.22.2
joblib>=1.4.0
multiprocess>=0.70.16
numpy>=1.26.4
openai>=1.23.3
pandas>=2.2.2
prompt-toolkit>=3.0.43
psutil>=5.9.8
pydantic>=2.7.0
pydantic-core>=2.18.1
pyyaml>=6.0.1
ray>=2.11.0
regex>=2024.4.16
requests>=2.31.0
scikit-learn>=1.4.2
scipy>=1.13.0
sympy>=1.12
termcolor>=2.4.0
tokenizers>=0.19.1
torch>=2.4.0
tqdm>=4.66.2
transformers
fschat>=0.2.36
peft>=0.11.1
tenacity>=8.4.1
vllm==0.8.5  
